+++ tests/testthat/test-time.R
@@ -137,6 +137,7 @@ test_with_dir("runtime predictions", {
 test_with_dir("predict_workers()", {
   skip_on_cran()
   skip_if_not_installed("knitr")
+  skip_if_not_installed("lubridate")
   load_mtcars_example()
   cache <- storr::storr_environment()
   config <- drake_config(my_plan, cache = cache, session_info = FALSE)